+++ middleware/handlePost.js
@@ -1,3 +1,5 @@
+const {postWinners, postLosers} = require('./crud-functions/index.js');
+
 async function handlePost(ctx) {
   const {channel_id} = ctx.request.body;
 }
